requirements.txt
openai
streamlit
langchain
requests
python-dotenv 
